+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - journaling: sauvegarde des données en base sur modification
 - journaling: changement du tri des données
 - journaling: ajout de la gestion via la console d'administration
+- journaling: refactoring de view.py
 
 ## [0.4.0]
 - documentaion: creation de l'application